+++ BambooTracker/bamboo_tracker.hpp
@@ -280,7 +280,7 @@ public:
 	std::string getStepEffectID(int songNum, int trackNum, int orderNum, int stepNum, int n) const;
 	void setStepEffectID(int songNum, int trackNum, int orderNum, int stepNum, int n, std::string id);
 	int getStepEffectValue(int songNum, int trackNum, int orderNum, int stepNum, int n) const;
-	void setStepEffectValue(int songNum, int trackNum, int orderNum, int stepNum, int n, int value);
+	void setStepEffectValue(int songNum, int trackNum, int orderNum, int stepNum, int n, int value, bool isFMReversed);
 	void eraseStepEffect(int songNum, int trackNum, int orderNum, int stepNum, int n);
 	void eraseStepEffectValue(int songNum, int trackNum, int orderNum, int stepNum, int n);
 	void deletePreviousStep(int songNum, int trackNum, int orderNum, int stepNum);