+++ BambooTracker/command/pattern/set_effect_value_to_step_command.hpp
@@ -7,7 +7,8 @@
 class SetEffectValueToStepCommand : public AbstractCommand
 {
 public:
-	SetEffectValueToStepCommand(std::weak_ptr<Module> mod, int songNum, int trackNum, int orderNum, int stepNum, int n, int value);
+	SetEffectValueToStepCommand(std::weak_ptr<Module> mod, int songNum, int trackNum,
+								int orderNum, int stepNum, int n, int value, bool isFMReversed);
 	void redo() override;
 	void undo() override;
 	int getID() const override;
@@ -25,4 +26,5 @@ private:
 	int song_, track_, order_, step_, n_;
 	int val_, prevVal_;
 	bool isComplete_;
+	bool isFMReserved_;
 };